+++ src/main/java/web/exception/LoginException.java
@@ -1,4 +1,4 @@
-package mvc.exception;
+package web.exception;
 
 public class LoginException extends RuntimeException {
     public LoginException() {
